removed unused deps, added business email to footer, restructured dirs
