+++ src/App.js
@@ -1,19 +1,16 @@
-import './App.css';
+import './css/App.css';
 import pfp from './images/pfp.png';
 import Projects from './components/Projects';
 import Socials from './components/Socials';
 import ParticlesBackground from './components/ParticlesBackground';
 import Contributions from './components/Contributions';
-import { useEffect, useState, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import Footer from './components/Footer';
 
 /*
 TODO:
 
-Change element IDs to classes
-Fill out footer with content
 Upload resume to S3
-Add email to footer
 */
 function App() {
   const galleryHead = useRef(null);
@@ -41,14 +38,14 @@ function App() {
     <div className="App">
     <ParticlesBackground/>
       <header className="App-header">
-        <div id="header-text">
-            <h1 id="main-header">NIKO RAISANEN</h1>
+        <div className="header-text">
+            <h1 className="main-header">NIKO RAISANEN</h1>
         </div>
 
-        <img src={pfp} id="header-pic" alt="headshot"/>
+        <img src={pfp} className="header-pic" alt="headshot"/>
         <Socials/>
         
-        <div id="intro">
+        <div className="intro">
             <p>Hello! I'm Niko, a software engineer with a background in security engineering. I have a passion for turning my ideas into tangible solutions through code and am always eager to dive into a new challenge.</p>
         </div>
       </header>
