+++ src/components/Footer.js
@@ -1,15 +1,22 @@
 import React from 'react';
-import './../footer.css';
+import '../css/footer.css';
 
 const Footer = () => {
   return (
 
     <footer>
-      <div class="footer-content">
+      <div className="footer-content">
         <h3>Download Resume</h3>
+        <a href="mailto:nikoraisanenbusiness@gmail.com" className='email-anchor' style={{'textDecoration': 'none'}}>
+          <p style={{'color': 'rgb(148, 145, 248)'}}>
+          <i className="fa fa-envelope" aria-hidden="true" style={{'paddingRight': '10px', 'color': 'white'}}></i>
+            nikoraisanenbusiness@gmail.com
+          </p>
+        </a>
+        
         <a href='#' className='download-anchor'>
           <svg width="35px" height="35px" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M20 15V18C20 19.1046 19.1046 20 18 20H6C4.89543 20 4 19.1046 4 18L4 15M8 11L12 15M12 15L16 11M12 15V3" stroke="#ffffff" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" className='download-icon'/>
+            <path d="M20 15V18C20 19.1046 19.1046 20 18 20H6C4.89543 20 4 19.1046 4 18L4 15M8 11L12 15M12 15L16 11M12 15V3" stroke="#ffffff" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" className='download-icon'/>
           </svg>
         </a>
         <p className='footer-text'>
